Move antlr extension into own file
