+++ crates/compiler/src/visitors.rs
@@ -3,7 +3,4 @@ mod declaration_visitor;
 mod last_line_before_options_visitor;
 mod string_table_generator_visitor;
 
-pub(crate) use self::{
-    constant_value_visitor::*, declaration_visitor::*, last_line_before_options_visitor::*,
-    string_table_generator_visitor::*,
-};
+pub(crate) use self::{last_line_before_options_visitor::*, string_table_generator_visitor::*};
